tests: split header check into non-Windows check
Including both the raylib header and windows.h (via perl.h) is problematic
because they define the same symbols. We might be able get around this by
defining NOUSER and friends to have the WinAPI not use these symbols:
https://github.com/raysan5/raylib/issues/857#issuecomment-498020922

But for now, just open code the prototype we need for our test and
have the header be used in a separate test that's skipped on Windows.

Fixes #4.